# Faster broadcast preprocess