--- a/pennylane/templates/broadcast.py
+++ b/pennylane/templates/broadcast.py
@@ -62,7 +62,29 @@ def wires_all_to_all(wires):
             sequence += [wires.subset([i, j])]
     return sequence
 
+# define wire sequences for patterns
+PATTERN_TO_WIRES = {
+    "single": lambda wires: [wires[i] for i in range(len(wires))],
+    "double": lambda wires: [wires.subset([i, i + 1]) for i in range(0, len(wires) - 1, 2)],
+    "double_odd": lambda wires: [wires.subset([i, i + 1]) for i in range(1, len(wires) - 1, 2)],
+    "chain": lambda wires: [wires.subset([i, i + 1]) for i in range(len(wires) - 1)],
+    "ring": lambda wires: wires_ring(wires),
+    "pyramid": lambda wires: wires_pyramid(wires),
+    "all_to_all": lambda wires: wires_all_to_all(wires),
+    "custom": lambda wires: wires,
+}
 
+# define required number of parameters
+PATTERN_TO_NUM_PARAMS = {
+    "single": lambda wires: len(wires),
+    "double": lambda wires: 0 if len(wires) in [0, 1] else len(wires) // 2,
+    "double_odd": lambda wires: 0 if len(wires) in [0, 1] else (len(wires) - 1) // 2,
+    "chain": lambda wires: 0 if len(wires) in [0, 1] else len(wires) - 1,
+    "ring": lambda wires: 0 if len(wires) in [0, 1] else (1 if len(wires) == 2 else len(wires)),
+    "pyramid": lambda wires: 0 if len(wires) in [0, 1] else sum(i + 1 for i in range(len(wires) // 2)),
+    "all_to_all": lambda wires: 0 if len(wires) in [0, 1] else len(wires) * (len(wires) - 1) // 2,
+    "custom": lambda wires: len(wires) if wires is not None else None,
+}
 ###################
 
 
@@ -82,40 +104,17 @@ def _preprocess(parameters, pattern, wires):
         wire_sequence, parameters: preprocessed pattern and parameters
     """
 
-    custom_pattern = None
 
     if isinstance(pattern, str):
+        _wires = wires
         if pattern not in OPTIONS:
             raise ValueError(f"did not recognize pattern {pattern}".format())
     else:
         # turn custom pattern into list of Wires objects
-        custom_pattern = [Wires(w) for w in pattern]
+        _wires = [Wires(w) for w in pattern]
         # set "pattern" to "custom", indicating that custom settings have to be used
         pattern = "custom"
 
-    # define wire sequences for patterns
-    pattern_to_wires = {
-        "single": lambda: [wires[i] for i in range(len(wires))],
-        "double": lambda: [wires.subset([i, i + 1]) for i in range(0, len(wires) - 1, 2)],
-        "double_odd": lambda: [wires.subset([i, i + 1]) for i in range(1, len(wires) - 1, 2)],
-        "chain": lambda: [wires.subset([i, i + 1]) for i in range(len(wires) - 1)],
-        "ring": lambda: wires_ring(wires),
-        "pyramid": lambda: wires_pyramid(wires),
-        "all_to_all": lambda: wires_all_to_all(wires),
-        "custom": lambda: custom_pattern,
-    }
-
-    # define required number of parameters
-    pattern_to_num_params = {
-        "single": lambda: len(wires),
-        "double": lambda: 0 if len(wires) in [0, 1] else len(wires) // 2,
-        "double_odd": lambda: 0 if len(wires) in [0, 1] else (len(wires) - 1) // 2,
-        "chain": lambda: 0 if len(wires) in [0, 1] else len(wires) - 1,
-        "ring": lambda: 0 if len(wires) in [0, 1] else (1 if len(wires) == 2 else len(wires)),
-        "pyramid": lambda: 0 if len(wires) in [0, 1] else sum(i + 1 for i in range(len(wires) // 2)),
-        "all_to_all": lambda: 0 if len(wires) in [0, 1] else len(wires) * (len(wires) - 1) // 2,
-        "custom": lambda: len(custom_pattern) if custom_pattern is not None else None,
-    }
 
     # check that there are enough parameters for pattern
     if parameters is not None:
@@ -139,15 +138,15 @@ def _preprocess(parameters, pattern, wires):
             raise ValueError(
                 "the ring pattern with 2 wires is an exception and only applies one unitary"
             )
-
-        if shape[0] != pattern_to_num_params[pattern]():
+        num_params = PATTERN_TO_NUM_PARAMS[pattern](_wires)
+        if shape[0] != num_params:
             raise ValueError(
                 "Parameters must contain entries for {} unitaries; got {} entries".format(
-                    pattern_to_num_params[pattern](), shape[0]
+                    num_params, shape[0]
                 )
             )
 
-    wire_sequence = pattern_to_wires[pattern]()
+    wire_sequence = PATTERN_TO_WIRES[pattern](_wires)
     return wire_sequence, parameters