**Context:**
Previously, the `_preprocess` would take upwards of 20% of the tape construction time because it was always building all of the wire patterns were always built with each function call. We remove this overhead, resulting in 15% speedup in tape construction, and ~2-3% e2e speedups for all interfaces.

Results from my laptop:

```
-      13.1±0.9ms       11.8±0.1ms     0.90  asv.core_suite.CircuitEvaluation.time_circuit(10, 3)
       23.6±0.3ms       21.8±0.5ms     0.93  asv.core_suite.CircuitEvaluation.time_circuit(10, 6)
       34.4±0.2ms       31.4±0.1ms     0.91  asv.core_suite.CircuitEvaluation.time_circuit(10, 9)
       3.68±0.2ms       3.48±0.2ms     0.95  asv.core_suite.CircuitEvaluation.time_circuit(2, 3)
       5.58±0.3ms       5.29±0.2ms     0.95  asv.core_suite.CircuitEvaluation.time_circuit(2, 6)
       7.46±0.3ms       7.24±0.3ms     0.97  asv.core_suite.CircuitEvaluation.time_circuit(2, 9)
       7.08±0.2ms       6.75±0.4ms     0.95  asv.core_suite.CircuitEvaluation.time_circuit(5, 3)
-      12.9±0.5ms      11.1±0.02ms     0.86  asv.core_suite.CircuitEvaluation.time_circuit(5, 6)
       17.2±0.3ms       16.0±0.3ms     0.93  asv.core_suite.CircuitEvaluation.time_circuit(5, 9)
       5.12±0.3ms       4.93±0.2ms     0.96  asv.core_suite.GradientComputation.time_gradient_autograd(2, 3)
       7.90±0.3ms       7.54±0.3ms     0.96  asv.core_suite.GradientComputation.time_gradient_autograd(2, 6)
       11.0±0.4ms       10.2±0.3ms     0.93  asv.core_suite.GradientComputation.time_gradient_autograd(5, 3)
       19.6±0.8ms         18.8±1ms     0.96  asv.core_suite.GradientComputation.time_gradient_autograd(5, 6)
       16.2±0.1ms       16.3±0.1ms     1.01  asv.core_suite.GradientComputation.time_gradient_tf(2, 3)
       28.1±0.4ms       28.8±0.6ms     1.03  asv.core_suite.GradientComputation.time_gradient_tf(2, 6)
       44.7±0.2ms       44.6±0.3ms     1.00  asv.core_suite.GradientComputation.time_gradient_tf(5, 3)
       88.1±0.5ms       87.3±0.4ms     0.99  asv.core_suite.GradientComputation.time_gradient_tf(5, 6)
       7.39±0.3ms       7.45±0.3ms     1.01  asv.core_suite.GradientComputation.time_gradient_torch(2, 3)
       19.5±0.6ms       18.6±0.9ms     0.95  asv.core_suite.GradientComputation.time_gradient_torch(2, 6)
       36.4±0.3ms       35.9±0.1ms     0.99  asv.core_suite.GradientComputation.time_gradient_torch(5, 3)
          129±2ms          126±2ms     0.98  asv.core_suite.GradientComputation.time_gradient_torch(5, 6)
```

**Description of the Change:**
Moved the wire pattern constructions to lambda methods, removing the overhead from all preprocess calls.

**Benefits:**
It's faster.

**Possible Drawbacks:**
None

**Related GitHub Issues:**
None